Add files via upload
Logistics is now a key factor in influencing the final cost of goods to the consumer. The role of logistics in business is difficult to overestimate. The visualization of data on the functioning of logistics supply chains greatly simplifies the work for all employees. From rank-and-file drivers to the entire department of a corporation's top managers.